display: move `Animation` to `components/animation`
